# ElectricCarsServer/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "App for serving autoparts requests",
        "title": "Autoparts App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/assembly": {
            "put": {
                "description": "Update Assembly by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Update Assembly by admin",
                "parameters": [
                    {
                        "description": "updated Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Assembly"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Assembly"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete Assembly by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Delete Assembly by admin",
                "parameters": [
                    {
                        "description": "Delete Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AssemblyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Assembly"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/assembly/complete": {
            "put": {
                "description": "Complete Assembly by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Complete Assembly by admin",
                "parameters": [
                    {
                        "description": "Complete Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AssemblyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Assembly"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/assembly/form": {
            "put": {
                "description": "Form Assembly by client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Form Assembly by client",
                "parameters": [
                    {
                        "description": "Form Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AssemblyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Assembly"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/assembly/get-al": {
            "get": {
                "description": "Assembly List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Assembly List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query string to filter Assemblies by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Query string to filter Assemblies from start date",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Query string to filter Assemblies to end date",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.Assembly"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/api/assembly/reject": {
            "put": {
                "description": "Reject Assembly by admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Reject Assembly by admin",
                "parameters": [
                    {
                        "description": "Reject Assembly",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AssemblyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Assembly"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/assembly{id}": {
            "get": {
                "description": "Assembly By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assembly"
                ],
                "summary": "Assembly By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Assembly ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.AssemblyDetails"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.Assembly": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "integer"
                },
                "date_end": {
                    "type": "string"
                },
                "date_processing": {
                    "type": "string"
                },
                "date_start": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "factory": {
                    "type": "string"
                },
                "factory_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "ds.AssemblyDetails": {
            "type": "object",
            "properties": {
                "assembly": {
                    "$ref": "#/definitions/ds.Assembly"
                },
                "autoparts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Autopart"
                    }
                }
            }
        },
        "ds.AssemblyForm": {
            "type": "object",
            "properties": {
                "factory_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "ds.Autopart": {
            "type": "object",
            "properties": {
                "autopart_id": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    }
}